+++ CysSubmission/CH.SC.U4CYS23051_golden_section_method.py
